Add files via upload
We will connect the five LEDs to pins 13, 12, 8, 4, and 2 of the Arduino board. The required resistance of the resistors is enough to light up an LED without damaging the board and other components.

The LED arranged one after another will light up. We can also change or rearrange the LEDs connected to the specified pin number on the board.